ДЗ 14 спринт: Логика работы плеера вынесена в отдельный класс. В активити осталась только работа с UI.
