+++ app/src/main/java/com/practicum/playlistmaker/activity/PlayerActivity.kt
@@ -5,8 +5,6 @@ import android.media.MediaPlayer
 import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
@@ -15,20 +13,17 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.model.Track
+import com.practicum.playlistmaker.model.TrackPlayer
 import com.practicum.playlistmaker.pixelConverter
 import com.practicum.playlistmaker.trackTimeFormat
-import java.text.SimpleDateFormat
-import java.util.Locale
 
 class PlayerActivity : AppCompatActivity() {
 
-    private lateinit var mainThreadHandler: Handler
     private lateinit var buttonPlayPause: ImageButton
-    private var mediaPlayer = MediaPlayer()
-    private var playerState = STATE_DEFAULT
     private var url: String? = null
-    private val timerUpdateRunnable = Runnable { updateTimer() }
     private lateinit var trackTimer: TextView
+    private var mediaPlayer = MediaPlayer()
+    private lateinit var trackPlayer: TrackPlayer
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -52,7 +47,7 @@ class PlayerActivity : AppCompatActivity() {
             intent.getParcelableExtra(TRACK_ID)
         }
 
-        url = track?.previewUrl
+        url = track?.previewUrl.toString()
 
         Glide
             .with(this)
@@ -69,72 +64,25 @@ class PlayerActivity : AppCompatActivity() {
         trackGenre.text = track.primaryGenreName
         trackCountry.text = track.country
 
-        mainThreadHandler = Handler(Looper.getMainLooper())
         buttonPlayPause = findViewById(R.id.button_play_pause)
-        preparePlayer()
-        buttonPlayPause.setOnClickListener { playbackControl() }
+        trackPlayer = TrackPlayer(this, mediaPlayer, url, buttonPlayPause, trackTimer)
+        trackPlayer.preparePlayer()
+        buttonPlayPause.setOnClickListener { trackPlayer.playbackControl() }
     }
 
     override fun onPause() {
         super.onPause()
-        pausePlayer()
-        mainThreadHandler.removeCallbacks(timerUpdateRunnable)
+        trackPlayer.pausePlayer()
+        trackPlayer.stopUpdater()
     }
 
     override fun onDestroy() {
         super.onDestroy()
         mediaPlayer.release()
-        mainThreadHandler.removeCallbacks(timerUpdateRunnable)
+        trackPlayer.stopUpdater()
     }
 
-    private fun preparePlayer(){
-        mediaPlayer.setDataSource(url)
-        mediaPlayer.prepareAsync()
-        mediaPlayer.setOnPreparedListener {
-            buttonPlayPause.isEnabled = true
-            playerState = STATE_PREPARED
-        }
-        mediaPlayer.setOnCompletionListener {
-            buttonPlayPause.setImageDrawable(getAttribute(R.attr.buttonPlay))
-            playerState = STATE_PREPARED
-            mainThreadHandler.removeCallbacks(timerUpdateRunnable)
-            trackTimer.text = ZERO_CONDITION
-        }
-    }
-
-    private fun playbackControl(){
-        when (playerState) {
-            STATE_PLAYING -> pausePlayer()
-            STATE_PREPARED, STATE_PAUSED -> startPlayer()
-        }
-    }
-
-    private fun startPlayer(){
-        mediaPlayer.start()
-        buttonPlayPause.setImageDrawable(getAttribute(R.attr.buttonPause))
-        playerState = STATE_PLAYING
-        mainThreadHandler.post(timerUpdateRunnable)
-    }
-
-    private fun pausePlayer(){
-        mediaPlayer.pause()
-        buttonPlayPause.setImageDrawable(getAttribute(R.attr.buttonPlay))
-        playerState = STATE_PAUSED
-    }
-
-    private fun updateTimer () : Runnable {
-        return object : Runnable {
-            override fun run() {
-                if (playerState == STATE_PLAYING){
-                    trackTimer.text = SimpleDateFormat("mm:ss", Locale.getDefault())
-                        .format(mediaPlayer.currentPosition)
-                    mainThreadHandler.postDelayed(this, DELAY)
-                }
-            }
-        }
-    }
-
-    private fun getAttribute(attr: Int): Drawable? {
+    fun getAttribute(attr: Int): Drawable? {
         val attrs = intArrayOf(attr)
         val typedArray = theme.obtainStyledAttributes(attrs)
         val drawableResourceId = typedArray.getResourceId(0, 0)
@@ -145,14 +93,6 @@ class PlayerActivity : AppCompatActivity() {
 
     companion object{
         const val TRACK_ID = "TRACK_ID"
-        const val ZERO_CONDITION = "00:00"
-
-        private const val STATE_DEFAULT = 0
-        private const val STATE_PREPARED = 1
-        private const val STATE_PLAYING = 2
-        private const val STATE_PAUSED = 3
-
-        private const val DELAY = 500L
     }
 }
