+++ app/src/main/java/com/practicum/playlistmaker/model/Debounce.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker
+package com.practicum.playlistmaker.model
 
 import android.content.Context
 import android.os.Handler
