+++ app/src/main/java/com/practicum/playlistmaker/searchHistory/SearchHistoryAdapter.kt
@@ -3,7 +3,7 @@ package com.practicum.playlistmaker.searchHistory
 import android.content.Intent
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.practicum.playlistmaker.Debounce
+import com.practicum.playlistmaker.model.Debounce
 import com.practicum.playlistmaker.activity.PlayerActivity
 import com.practicum.playlistmaker.model.Track
 import com.practicum.playlistmaker.trackList.TrackListViewHolder
